[#30] [#26] Code clean up and refactoring the functionality feature, the dock color can be customized and saved after application shutdown
