+++ app/src/main/kotlin/com/simplemobiletools/camera/helpers/MyPreference.kt
@@ -1,4 +1,4 @@
-package com.simplemobiletools.camera.activities
+package com.simplemobiletools.camera.helpers
 
 import android.content.Context
 
@@ -8,11 +8,11 @@ class MyPreference(context: Context){
 
     val preference = context.getSharedPreferences(PREFERENCE_NAME, Context.MODE_PRIVATE)
 
-    fun getLoginCount() : Int {
+    fun getDockerColor() : Int {
         return preference.getInt(PREFERENCE_COLOR, -7508381)
     }
 
-    fun setLoginCount(color:Int){
+    fun setDockerColor(color:Int){
         val editor = preference.edit()
         editor.putInt(PREFERENCE_COLOR, color)
         editor.apply()
